changed zuBestellendeGebinde to double

+++ domain/src/main/java/de/dhbw/foodcoop/warehouse/domain/entities/DiscrepancyEntity.java
@@ -20,7 +20,7 @@ public class DiscrepancyEntity {
 	
 	// Wenn ein Gebinde z.B. 2,5kg ist, und zuBestellendeGebinde 3, heißt das 2,5 * 3 für gesamte Kg Menge
 	@Column
-	private int zuBestellendeGebinde;
+	private double zuBestellendeGebinde;
 	
 	// Positive Zahl: Menge die zu viel geliefert wurde. In Stückzahl oder Kg
 	// Negative Zahl: Menge die zu wenig geliefert wurde. In Stückzahl oder Kg
@@ -33,7 +33,7 @@ public class DiscrepancyEntity {
 	
 
 	
-	public DiscrepancyEntity(String id, BestandEntity bestand, int zuBestellendeGebinde, float zuVielzuWenig, float gewollteMenge) {
+	public DiscrepancyEntity(String id, BestandEntity bestand, double zuBestellendeGebinde, float zuVielzuWenig, float gewollteMenge) {
 		super();
 		this.gewollteMenge = gewollteMenge;
 		this.id = id;
@@ -58,11 +58,11 @@ public class DiscrepancyEntity {
 
 
 
-	public int getZuBestellendeGebinde() {
+	public double getZuBestellendeGebinde() {
 		return zuBestellendeGebinde;
 	}
 
-	public void setZuBestellendeGebinde(int zuBestellendeGebinde) {
+	public void setZuBestellendeGebinde(double zuBestellendeGebinde) {
 		this.zuBestellendeGebinde = zuBestellendeGebinde;
 	}
